Add custom FormField to select movies.
MovieTitleField is a custom form field that allows the CMS user to
search for a movie using a TextField which then updates a Dropdown field
based on the results returned from the controller.

Makes use of the MovieInformationController action getmovies which in
turn uses the OMDb API to search based on a query.

A javascript library is used to talk to the controller and update the
dropdown field.