Add frontend to MovieInformation page
Allows a public user to view information about a saved MovieInformtion
page. Adds a new controller function to get information about the movie
including rotten tomato information as well as both short and long
version of the plot.

This information is retrieved from the OMDb API.

Adds both the template and the CSS for this display.